reduce mem usage

+++ code/models/base_connector.py
@@ -370,7 +370,7 @@ class Base_Connector():
 
     def save_fc1_act_hook(self, layer) -> Callable:
         def fn(_, __, output):
-            self.fc1_output[layer] = output.detach().cpu()
+            self.fc1_output[layer] = output.detach().cpu().half()
         return fn
 
     def get_fc1_act(self):
